+++ rnd/autogpt_builder/src/api/autogpt_server_api.ts
@@ -1,5 +1,4 @@
-import { XYPosition } from "reactflow";
-import { ObjectSchema } from "./types";
+import {Block, NodeExecutionResult} from "@/types/api";
 
 export default class AutoGPTServerAPI {
   private baseUrl: string;
@@ -127,67 +126,4 @@ export default class AutoGPTServerAPI {
   }
 }
 
-/* Mirror of autogpt_server/data/block.py:Block */
-export type Block = {
-  id: string;
-  name: string;
-  description: string;
-  inputSchema: ObjectSchema;
-  outputSchema: ObjectSchema;
-};
 
-/* Mirror of autogpt_server/data/graph.py:Node */
-export type Node = {
-  id: string;
-  block_id: string;
-  input_default: Map<string, any>;
-  input_nodes: Array<{ name: string, node_id: string }>;
-  output_nodes: Array<{ name: string, node_id: string }>;
-  metadata: {
-    position: XYPosition;
-    [key: string]: any;
-  };
-};
-
-/* Mirror of autogpt_server/data/graph.py:Link */
-export type Link = {
-  source_id: string;
-  sink_id: string;
-  source_name: string;
-  sink_name: string;
-}
-
-/* Mirror of autogpt_server/data/graph.py:Graph */
-export type Flow = {
-  id: string;
-  name: string;
-  description: string;
-  nodes: Array<Node>;
-  links: Array<Link>;
-};
-
-export type FlowCreateBody = Flow | {
-  id?: string;
-}
-
-/* Derived from autogpt_server/executor/manager.py:ExecutionManager.add_execution */
-export type FlowExecuteResponse = {
-  /* ID of the initiated run */
-  id: string;
-  /* List of node executions */
-  executions: Array<{ id: string, node_id: string }>;
-};
-
-/* Mirror of autogpt_server/data/execution.py:ExecutionResult */
-export type NodeExecutionResult = {
-  graph_exec_id: string;
-  node_exec_id: string;
-  node_id: string;
-  status: 'INCOMPLETE' | 'QUEUED' | 'RUNNING' | 'COMPLETED' | 'FAILED';
-  input_data: Map<string, any>;
-  output_data: Map<string, any[]>;
-  add_time: Date;
-  queue_time?: Date;
-  start_time?: Date;
-  end_time?: Date;
-};
