--- rnd/autogpt_builder/src/lib/types.ts
@@ -1,6 +1,0 @@
-export type ObjectSchema = {
-    type: string;
-    properties: { [key: string]: any };
-    additionalProperties?: { type: string };
-    required?: string[];
-  };
